refactor(domain): :art: refactor repository files
